+++ tests/back/Channel/Specification/Bundle/Doctrine/Remover/ChannelRemoverSpec.php
@@ -3,6 +3,7 @@
 namespace Specification\Akeneo\Channel\Bundle\Doctrine\Remover;
 
 use Akeneo\Channel\Component\Model\Channel;
+use Akeneo\Channel\Component\Query\IsChannelUsedInProductExportJobInterface;
 use Akeneo\Tool\Component\StorageUtils\Remover\RemoverInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
@@ -16,13 +17,15 @@ class ChannelRemoverSpec extends ObjectBehavior
         ObjectManager $objectManager,
         EventDispatcherInterface $eventDispatcher,
         ChannelRepositoryInterface $channelRepository,
-        TranslatorInterface $translator
+        TranslatorInterface $translator,
+        IsChannelUsedInProductExportJobInterface $isChannelUsedInProductExportJob
     ) {
         $this->beConstructedWith(
             $objectManager,
             $eventDispatcher,
             $channelRepository,
             $translator,
+            $isChannelUsedInProductExportJob,
             Channel::class
         );
     }
@@ -35,9 +38,15 @@ class ChannelRemoverSpec extends ObjectBehavior
     function it_removes_the_channel_and_flushes_the_unit_of_work(
         $objectManager,
         Channel $channel,
-        ChannelRepositoryInterface $channelRepository
+        ChannelRepositoryInterface $channelRepository,
+        IsChannelUsedInProductExportJobInterface $isChannelUsedInProductExportJob
     ) {
+        $channel->getId()->willReturn(42);
+        $channel->getCode()->willReturn('mobile');
+
         $channelRepository->countAll()->willReturn(2);
+        $isChannelUsedInProductExportJob->execute('mobile')->willReturn(false);
+
         $objectManager->remove($channel)->shouldBeCalled();
         $objectManager->flush()->shouldBeCalled();
         $this->remove($channel);
@@ -64,6 +73,23 @@ class ChannelRemoverSpec extends ObjectBehavior
         $channelRepository->countAll()->willReturn(1);
         $channel->getCode()->willReturn('code');
         $translator->trans('pim_enrich.channel.flash.delete.error', ['%channelCode%' => 'code'])->willReturn('exception message');
+        $logicException = new \LogicException('exception message');
+
+        $this->shouldThrow($logicException)->during('remove', [$channel]);
+    }
+
+    public function it_throws_logic_exception_when_the_channel_is_used_in_an_export_profile(
+        Channel $channel,
+        ChannelRepositoryInterface $channelRepository,
+        TranslatorInterface $translator,
+        IsChannelUsedInProductExportJobInterface $isChannelUsedInProductExportJob
+    ) {
+        $channel->getCode()->willReturn('mobile');
+
+        $channelRepository->countAll()->willReturn(2);
+        $isChannelUsedInProductExportJob->execute('mobile')->willReturn(true);
+
+        $translator->trans('pim_enrich.channel.flash.delete.linked_to_export_profile')->willReturn('exception message');
         $logicException = new \LogicException('exception message');
 
         $this->shouldThrow($logicException)->during('remove', [$channel]);